Update Chip Tray Photo Processor.py
Added dark / light mode because why not.

Changed the GUI around a bit hopefully to make it a bit more clear.... need critique on it though cos i haven't a clue.

Added Full French translation (Language button in the top left)

Fixed a hundred different minor things which weren't working. Broke some other stuff in the mean time.

QAQC photo review is very rewarding.

added an HEIC --> JPG/PNG converter into the tools

going to have to keep on top of the people on site to ensure that they're checking the photos they're taking - had several we've had to reject due to quality.